# osdu-r2/os-delivery/docs/delivery-service_openapi.yaml
# Copyright 2019 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

openapi: 3.0.0

info:
  version: "0.0.1"
  title: OpenDES Delivery Service
  description: |
    The OpenDES Delivery service provides internal and external APIs to request file location data, such as an individual signed URL per file.

servers:
  - url: amer-demo28-test.apigee.net/os-file
    description: Test server

paths:
  /getLocation:
    post:
      description: |
        Create a new location in the landing zone, such as a GCS bucket.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Partition'
      requestBody:
        $ref: '#/components/requestBodies/LocationRequest'
      responses:
        '200':
          description: |
            Created location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/Bad-Request'
        '5XX':
          description: Unexpected error.
  /getFileLocation:
    post:
      description: |
        Returns file `Location` and `Driver`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Partition'
      requestBody:
        $ref: '#/components/requestBodies/FileLocationRequest'
      responses:
        '200':
          description: |
            File location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileLocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/Bad-Request'
        '5XX':
          description: Unexpected error.
  /getFileList:
    post:
      description: |
        Allows the application to audit the attempted file uploads. The method is internal and isn't available for third-party applications.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Partition'
      requestBody:
        $ref: '#/components/requestBodies/FileListRequest'
      responses:
        '200':
          description: |
            File list page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/Bad-Request'
        '5XX':
          description: Unexpected error.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Partition:
      in: header
      name: Partition-Id
      description: |
        A unique identifier of a tenant
      required: true
      schema:
        type: string
  schemas:
    DateTime:
      type: string
      format: date-time
    FileID:
      type: string
      pattern: ^[\w,\s-]+(\.\w+)?$
    Driver:
      type: string
      enum:
        - GCP
        - AWS
    FileLocation:
      type: object
      properties:
        FileID:
          $ref: '#/components/schemas/FileID'
        Driver:
          $ref: '#/components/schemas/Driver'
        Location:
          type: string
        CreatedAt:
          type: string
          format: date-time
        CreatedBy:
          type: string
    ApiError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    LocationResponse:
      type: object
      properties:
        FileID:
          $ref: '#/components/schemas/FileID'
        Location:
          type: object
          additionalProperties:
            type: string
          example:
            SignedURL: 'GCS signed URL'
    FileLocationResponse:
      type: object
      properties:
        Driver:
          $ref: '#/components/schemas/Driver'
        Location:
          type: string
    FileListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FileLocation'
        number:
          type: integer
        numberOfElements:
          type: integer
        size:
          type: integer

  requestBodies:
    LocationRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              FileID:
                $ref: '#/components/schemas/FileID'
    FileLocationRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              FileID:
                $ref: '#/components/schemas/FileID'
            required:
              - FileID
    FileListRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              TimeFrom:
                $ref: '#/components/schemas/DateTime'
              TimeTo:
                $ref: '#/components/schemas/DateTime'
              PageNum:
                type: integer
              Items:
                type: integer
              UserID:
                type: string
  responses:
    Unauthorized:
      description: Unauthorized
    Bad-Request:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'